# docs/openapi/rag-service.json
{"openapi":"3.1.0","info":{"title":"NIA RAG Service","description":"**[Skill]** Retrieval-Augmented Generation for tenant knowledge bases. Ingest Markdown, plain text or JSON documents into Qdrant, then answer questions grounded in the retrieved chunks via the configured LLM. Invoked by the orchestrator when the FSM routes to the `faq` action.","version":"1.0.0"},"paths":{"/health/ready":{"get":{"tags":["ops"],"summary":"Readiness","operationId":"readiness_health_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/rag/query":{"post":{"tags":["rag"],"summary":"Query the RAG knowledge base","description":"Query the knowledge base for a tenant and return an LLM-generated answer\ngrounded in the retrieved chunks.\n\n`collection_name` defaults to `{tenant_id}_docs` if not provided.","operationId":"rag_query_v1_rag_query_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_RAGQueryResult_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/rag/ingest":{"post":{"tags":["rag"],"summary":"Ingest a document into the knowledge base","description":"Ingest a document into the tenant knowledge base.\n\n- **Plain text / Markdown** (`.txt`, `.md`): chunked by paragraph and embedded.\n- **JSON file** (`.json`): auto-detected. Each top-level object becomes one chunk.\n  Supports a root array `[{...}]` or a dict with one list value `{\"items\": [{...}]}`.\n\n`collection_name` defaults to `{tenant_id}_docs`.","operationId":"ingest_v1_rag_ingest_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_ingest_v1_rag_ingest_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_IngestResponse_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/rag/documents/{doc_id}":{"delete":{"tags":["rag"],"summary":"Delete all chunks for a document","description":"Delete all vector chunks associated with a specific document ID.\n`collection_name` defaults to `{tenant_id}_docs`.","operationId":"delete_document_v1_rag_documents__doc_id__delete","parameters":[{"name":"doc_id","in":"path","required":true,"schema":{"type":"string","title":"Doc Id"}},{"name":"tenant_id","in":"query","required":true,"schema":{"type":"string","title":"Tenant Id"}},{"name":"collection_name","in":"query","required":false,"schema":{"type":"string","default":"","title":"Collection Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIResponse_DeleteResponse_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"APIResponse_DeleteResponse_":{"properties":{"data":{"$ref":"#/components/schemas/DeleteResponse"},"meta":{"$ref":"#/components/schemas/ResponseMeta"},"_links":{"additionalProperties":{"$ref":"#/components/schemas/HATEOASLink"},"type":"object","title":" Links","default":{}}},"type":"object","required":["data"],"title":"APIResponse[DeleteResponse]"},"APIResponse_IngestResponse_":{"properties":{"data":{"$ref":"#/components/schemas/IngestResponse"},"meta":{"$ref":"#/components/schemas/ResponseMeta"},"_links":{"additionalProperties":{"$ref":"#/components/schemas/HATEOASLink"},"type":"object","title":" Links","default":{}}},"type":"object","required":["data"],"title":"APIResponse[IngestResponse]"},"APIResponse_RAGQueryResult_":{"properties":{"data":{"$ref":"#/components/schemas/RAGQueryResult"},"meta":{"$ref":"#/components/schemas/ResponseMeta"},"_links":{"additionalProperties":{"$ref":"#/components/schemas/HATEOASLink"},"type":"object","title":" Links","default":{}}},"type":"object","required":["data"],"title":"APIResponse[RAGQueryResult]"},"Body_ingest_v1_rag_ingest_post":{"properties":{"tenant_id":{"type":"string","title":"Tenant Id"},"file":{"type":"string","format":"binary","title":"File"},"collection_name":{"type":"string","title":"Collection Name","default":""}},"type":"object","required":["tenant_id","file"],"title":"Body_ingest_v1_rag_ingest_post"},"DeleteResponse":{"properties":{"deleted":{"type":"boolean","title":"Deleted"},"doc_id":{"type":"string","title":"Doc Id"},"collection_name":{"type":"string","title":"Collection Name"}},"type":"object","required":["deleted","doc_id","collection_name"],"title":"DeleteResponse"},"HATEOASLink":{"properties":{"href":{"type":"string","title":"Href"},"method":{"type":"string","title":"Method","default":"GET"},"rel":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Rel"}},"type":"object","required":["href"],"title":"HATEOASLink"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IngestResponse":{"properties":{"doc_id":{"type":"string","title":"Doc Id"},"filename":{"type":"string","title":"Filename"},"chunks_created":{"type":"integer","title":"Chunks Created"},"tenant_id":{"type":"string","title":"Tenant Id"},"collection_name":{"type":"string","title":"Collection Name"}},"type":"object","required":["doc_id","filename","chunks_created","tenant_id","collection_name"],"title":"IngestResponse"},"QueryRequest":{"properties":{"query":{"type":"string","title":"Query"},"tenant_id":{"type":"string","title":"Tenant Id"},"collection_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Collection Name"},"tenant_name":{"type":"string","title":"Tenant Name","default":"el asistente"}},"type":"object","required":["query","tenant_id"],"title":"QueryRequest","example":{"query":"¿Cuál es el horario de la viña?","tenant_id":"demo_turismo"}},"RAGChunkSource":{"properties":{"doc_id":{"type":"string","title":"Doc Id"},"doc_name":{"type":"string","title":"Doc Name"},"chunk_index":{"type":"integer","title":"Chunk Index"},"section":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Section"},"retrieval_score":{"type":"number","title":"Retrieval Score"},"rerank_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rerank Score"},"text_excerpt":{"type":"string","title":"Text Excerpt"}},"type":"object","required":["doc_id","doc_name","chunk_index","retrieval_score","text_excerpt"],"title":"RAGChunkSource"},"RAGQueryResult":{"properties":{"query":{"type":"string","title":"Query"},"answer":{"type":"string","title":"Answer"},"confidence_score":{"type":"number","title":"Confidence Score"},"chunks_used":{"items":{"$ref":"#/components/schemas/RAGChunkSource"},"type":"array","title":"Chunks Used"},"groundedness_check":{"type":"string","title":"Groundedness Check","default":"passed"},"retrieval_model":{"type":"string","title":"Retrieval Model"},"generation_model":{"type":"string","title":"Generation Model"},"latency_retrieval_ms":{"type":"number","title":"Latency Retrieval Ms"},"latency_generation_ms":{"type":"number","title":"Latency Generation Ms"},"is_fallback":{"type":"boolean","title":"Is Fallback","default":false}},"type":"object","required":["query","answer","confidence_score","chunks_used","retrieval_model","generation_model","latency_retrieval_ms","latency_generation_ms"],"title":"RAGQueryResult"},"ResponseMeta":{"properties":{"request_id":{"type":"string","title":"Request Id"},"timestamp":{"type":"string","title":"Timestamp"},"version":{"type":"string","title":"Version","default":"1"}},"type":"object","title":"ResponseMeta"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"rag","description":"Document ingestion, vector search and LLM-grounded answers."},{"name":"ops","description":"Health and readiness probes."}]}